refactor(qa): split desktop god objects and harden runtime dependencies
- extract desktop direct chat, company event stream, issue lifecycle, Swift inspector, operator chat, dashboard merge, and text helper boundaries

- harden shell wrapper build lock recovery and safe jar discovery

- centralize direct chat provider catalog and stabilize local Playwright dependency resolution

- preserve completed QA review task reconciliation before stale workflow lineage repair rebuilds review or approval issues

Tested: ./gradlew --no-daemon formatCheck --console=plain; ./gradlew test --no-daemon --console=plain; swift test; swift build; ./shell/cotor version; bash -n shell/cotor; bash -n shell/build-desktop-app-bundle.sh; graphify update .; git diff --check; isolated app-server /health smoke

Confidence: high
