+++ macos/Sources/CotorDesktopApp/ContentView.swift
@@ -7608,293 +7608,6 @@ private func usdCurrencyText(_ cents: Int?) -> String {
     return formatter.string(from: NSNumber(value: Double(cents) / 100.0)) ?? String(format: "$%.2f", Double(cents) / 100.0)
 }
 
-private func userFacingPathLabel(_ path: String) -> String {
-    let trimmed = path.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    if trimmed.contains("://") {
-        return trimmed
-    }
-    let name = URL(fileURLWithPath: trimmed).lastPathComponent
-    return name.isEmpty ? trimmed : name
-}
-
-private func userFacingIssueTitle(_ title: String, language: AppLanguage) -> String {
-    let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-
-    if trimmed.hasPrefix("QA review ") {
-        let nested = userFacingIssueTitle(String(trimmed.dropFirst("QA review ".count)), language: language)
-        return language("QA review · \(nested)", "QA 검토 · \(nested)")
-    }
-    if trimmed.hasPrefix("CEO approve ") {
-        let nested = userFacingIssueTitle(String(trimmed.dropFirst("CEO approve ".count)), language: language)
-        return language("CEO approval · \(nested)", "CEO 확인 · \(nested)")
-    }
-    if trimmed.hasPrefix("CEO plan and delegate ") {
-        let goal = quotedGoal(in: trimmed) ?? trailingGoal(after: " ", in: trimmed)
-        return issueTitle("CEO planning", "CEO 작업 분배", goal: goal, language: language)
-    }
-    if trimmed.hasPrefix("Restore GitHub publishing for ") {
-        let nested = userFacingIssueTitle(
-            String(trimmed.dropFirst("Restore GitHub publishing for ".count)),
-            language: language
-        )
-        return language("Connect GitHub · \(nested)", "GitHub 연결 설정 · \(nested)")
-    }
-
-    let goal = quotedGoal(in: trimmed) ?? trailingGoal(after: " for ", in: trimmed)
-    if trimmed.hasPrefix("Deliver the first implementation slice for ") ||
-        trimmed.hasPrefix("Build the first working change for ") {
-        return issueTitle("First working change", "첫 작업 만들기", goal: goal, language: language)
-    }
-    if trimmed.hasPrefix("Advance a second branchable improvement slice for ") ||
-        trimmed.hasPrefix("Improve another small part of ") {
-        return issueTitle("Improve another part", "다음 부분 개선", goal: goal, language: language)
-    }
-    if trimmed.hasPrefix("Harden the integration and failure-handling path for ") ||
-        (trimmed.hasPrefix("Make ") && trimmed.hasSuffix(" more reliable")) {
-        return issueTitle("Improve reliability", "안정성 개선", goal: goal, language: language)
-    }
-    if trimmed.hasPrefix("Prepare validation evidence and residual-risk callouts for ") ||
-        trimmed.hasPrefix("Check the result and write down any remaining risks for ") {
-        return issueTitle("Check results and risks", "결과와 위험 확인", goal: goal, language: language)
-    }
-    return trimmed
-}
-
-private func userFacingMemoryEvent(_ event: String, language: AppLanguage) -> String {
-    let trimmed = event.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    if trimmed.hasPrefix("Started issue run: ") {
-        let title = userFacingIssueTitle(String(trimmed.dropFirst("Started issue run: ".count)), language: language)
-        return language("Started: \(title)", "시작: \(title)")
-    }
-    if trimmed.hasPrefix("Delegated issue: ") {
-        let remainder = String(trimmed.dropFirst("Delegated issue: ".count))
-        let title = remainder.components(separatedBy: " -> ").first ?? remainder
-        return language(
-            "Assigned: \(userFacingIssueTitle(title, language: language))",
-            "배정: \(userFacingIssueTitle(title, language: language))"
-        )
-    }
-    if trimmed.hasPrefix("Created infra issue: ") {
-        let title = String(trimmed.dropFirst("Created infra issue: ".count))
-        return language(
-            "Created connection task: \(userFacingIssueTitle(title, language: language))",
-            "연결 작업 생성: \(userFacingIssueTitle(title, language: language))"
-        )
-    }
-    if trimmed.hasPrefix("Blocked code issue: ") {
-        let reason = String(trimmed.dropFirst("Blocked code issue: ".count))
-        return language(
-            "Code work blocked: \(userFacingIssueDescription(reason, language: language))",
-            "코드 작업 차단: \(userFacingIssueDescription(reason, language: language))"
-        )
-    }
-    return userFacingIssueTitle(trimmed, language: language)
-}
-
-private func userFacingIssueDescription(_ description: String, language: AppLanguage) -> String {
-    let trimmed = description.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-
-    if trimmed.hasPrefix("CEO planning issue for the company goal.") {
-        return language(
-            "CEO is turning the goal into team work.",
-            "CEO가 목표를 팀 작업으로 나누는 준비입니다."
-        )
-    }
-
-    if trimmed.hasPrefix("QA review issue for a concrete pull request.") {
-        if let executionIssue = labeledLineValue("Execution issue:", in: trimmed) {
-            let title = userFacingIssueTitle(executionIssue, language: language)
-            return language("QA is reviewing \(title).", "QA 검토 중: \(title)")
-        }
-        return language("QA is reviewing the finished work.", "QA가 완료된 작업을 검토합니다.")
-    }
-
-    if trimmed.contains("GitHub publishing is required before this code issue can continue.") ||
-        trimmed.contains("GitHub PR mode requires an existing origin remote.") ||
-        trimmed.contains("GitHub is not connected.") {
-        return language(
-            "Connect an existing GitHub repository before starting PR-based code work. Cotor will not create one automatically.",
-            "PR 기반 코드 작업을 시작하려면 기존 GitHub 저장소를 먼저 연결해야 합니다. Cotor가 저장소를 자동으로 만들지는 않습니다."
-        )
-    }
-
-    if trimmed.contains("Owned subtasks:") || trimmed.contains("Shared checklist:") || trimmed.hasPrefix("Goal:") {
-        let role = labeledLineValue("Role:", in: trimmed)
-        let phase = labeledLineValue("Phase:", in: trimmed)
-        let focus = labeledLineValue("Focus:", in: trimmed)
-
-        var parts: [String] = []
-        if let role {
-            parts.append(language("\(role) owns this work.", "\(role)이(가) 맡은 작업입니다."))
-        }
-        if let phase {
-            parts.append(language("Stage: \(plainPhaseLabel(phase, language: language)).", "단계: \(plainPhaseLabel(phase, language: language))."))
-        }
-        if let focus, !containsInternalMemoryDetail(focus) {
-            parts.append(language("Focus is already set.", "작업 방향이 정해져 있습니다."))
-        }
-        return parts.isEmpty
-            ? language("Assigned team work with checks prepared.", "검증까지 준비된 팀 작업입니다.")
-            : parts.joined(separator: " ")
-    }
-
-    if containsInternalMemoryDetail(trimmed) {
-        return language("Details available in the issue record.", "세부 정보는 이슈 기록에서 확인할 수 있습니다.")
-    }
-    return trimmed
-}
-
-private func userFacingIssueKind(_ kind: String, language: AppLanguage) -> String {
-    switch kind.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() {
-    case "code":
-        return language("Code work", "코드 작업")
-    case "infra":
-        return language("Connection", "연결 문제")
-    case "execution":
-        return language("Team work", "팀 작업")
-    case "planning":
-        return language("Planning", "계획")
-    case "review":
-        return language("Review", "검토")
-    case "approval":
-        return language("Approval", "승인")
-    default:
-        return kind.isEmpty ? "—" : kind
-    }
-}
-
-private func userFacingDecisionTitle(_ title: String, language: AppLanguage) -> String {
-    let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    if trimmed == "Fallback planned execution graph" {
-        return language("Fallback execution plan", "기본 실행 계획")
-    }
-    return trimmed
-}
-
-private func userFacingDecisionSummary(_ summary: String, language: AppLanguage) -> String {
-    let trimmed = summary.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    if trimmed.contains("Fallback planner decomposed") {
-        return language(
-            "CEO planning needed a fallback, so Cotor split the goal into clear team issues.",
-            "CEO 계획이 막혀 기본 계획으로 목표를 팀 이슈로 나눴습니다."
-        )
-    }
-    if containsInternalMemoryDetail(trimmed) {
-        return language("Details available in the issue record.", "세부 정보는 이슈 기록에서 확인할 수 있습니다.")
-    }
-    return trimmed
-}
-
-private func userFacingActivityTitle(_ title: String, language: AppLanguage) -> String {
-    let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    switch trimmed {
-    case "Created infra issue":
-        return language("Created connection task", "연결 작업 생성")
-    case "Blocked code issue":
-        return language("Code work blocked", "코드 작업 차단")
-    case "Unblocked code issue":
-        return language("Code work unblocked", "코드 작업 재개")
-    case "Resolved infra issue":
-        return language("Connection task resolved", "연결 작업 해결")
-    default:
-        return userFacingIssueTitle(trimmed, language: language)
-    }
-}
-
-private func userFacingActivityDetail(_ detail: String, language: AppLanguage) -> String {
-    let trimmed = detail.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "—" }
-    if trimmed.hasPrefix("Restore GitHub publishing for ") {
-        return userFacingIssueTitle(trimmed, language: language)
-    }
-    if trimmed.contains("GitHub publishing is required before this code issue can continue.") ||
-        trimmed.contains("GitHub is not connected.") ||
-        trimmed.contains("GitHub PR mode requires an existing origin remote.") {
-        return userFacingIssueDescription(trimmed, language: language)
-    }
-    if trimmed.contains("Fallback planner decomposed") {
-        return userFacingDecisionSummary(trimmed, language: language)
-    }
-    if trimmed.contains("Owned subtasks:") || trimmed.hasPrefix("Goal:") || trimmed.hasPrefix("QA review issue for") {
-        return userFacingIssueDescription(trimmed, language: language)
-    }
-    if containsInternalMemoryDetail(trimmed) {
-        return language("Details available in the issue record.", "세부 정보는 이슈 기록에서 확인할 수 있습니다.")
-    }
-    return trimmed
-}
-
-private func containsInternalMemoryDetail(_ value: String) -> Bool {
-    let lowered = value.lowercased()
-    return lowered.contains("/users/")
-        || lowered.contains("graphify ")
-        || lowered.contains("graphify=")
-        || lowered.contains(".json")
-}
-
-private func labeledLineValue(_ label: String, in value: String) -> String? {
-    let lines = value.split(whereSeparator: \.isNewline).map(String.init)
-    for (index, line) in lines.enumerated() {
-        let trimmed = line.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.hasPrefix(label) {
-            let inline = trimmed.dropFirst(label.count).trimmingCharacters(in: .whitespacesAndNewlines)
-            if !inline.isEmpty {
-                return inline
-            }
-            let nextIndex = index + 1
-            guard nextIndex < lines.count else { return nil }
-            let next = lines[nextIndex].trimmingCharacters(in: .whitespacesAndNewlines)
-            return next.isEmpty ? nil : next
-        }
-    }
-    return nil
-}
-
-private func plainPhaseLabel(_ phase: String, language: AppLanguage) -> String {
-    switch phase.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() {
-    case "planning":
-        return language("planning", "계획")
-    case "execution":
-        return language("execution", "실행")
-    case "review":
-        return language("review", "검토")
-    case "approval":
-        return language("approval", "승인")
-    default:
-        return phase
-    }
-}
-
-private func quotedGoal(in value: String) -> String? {
-    guard let firstQuote = value.firstIndex(of: "\"") else { return nil }
-    let afterFirst = value.index(after: firstQuote)
-    guard let secondQuote = value[afterFirst...].firstIndex(of: "\"") else { return nil }
-    let goal = String(value[afterFirst..<secondQuote]).trimmingCharacters(in: .whitespacesAndNewlines)
-    return goal.isEmpty ? nil : goal
-}
-
-private func trailingGoal(after marker: String, in value: String) -> String? {
-    guard let markerRange = value.range(of: marker, options: [.backwards]) else { return nil }
-    let goal = value[markerRange.upperBound...]
-        .trimmingCharacters(in: .whitespacesAndNewlines)
-        .trimmingCharacters(in: CharacterSet(charactersIn: "\""))
-    return goal.isEmpty ? nil : goal
-}
-
-private func issueTitle(_ english: String, _ korean: String, goal: String?, language: AppLanguage) -> String {
-    guard let goal, !goal.isEmpty else {
-        return language(english, korean)
-    }
-    return language("\(english) · \(goal)", "\(korean) · \(goal)")
-}
-
 private func relativeTimestamp(_ value: Int64) -> String {
     guard value > 0 else { return "—" }
     let formatter = RelativeDateTimeFormatter()
@@ -8161,377 +7874,6 @@ private struct RunTabButton: View {
             .contentShape(Rectangle())
         }
         .buttonStyle(.plain)
-    }
-}
-
-private struct InspectorPaneView: View {
-    @EnvironmentObject private var store: DesktopStore
-    var embedded: Bool = false
-    private var l: AppLanguage { store.language }
-
-    var body: some View {
-        Group {
-            inspectorContent
-        }
-        .modifier(InspectorContainerModifier(embedded: embedded))
-    }
-
-    private var inspectorContent: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack(alignment: .top) {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(store.selectedIssue?.title ?? store.selectedRun?.agentName ?? l.text(.inspect))
-                        .font(.system(size: 20, weight: .semibold))
-                        .foregroundStyle(ShellPalette.text)
-                    Text(
-                        store.selectedReviewQueueItem.map { l.status($0.status) }
-                            ?? store.selectedTask?.title
-                            ?? l.text(.inspectorSubtitle)
-                    )
-                        .font(.system(size: 12, weight: .medium))
-                        .foregroundStyle(ShellPalette.muted)
-                        .lineLimit(2)
-                }
-
-                Spacer()
-
-                if let selectedTask = store.selectedTask, !selectedTask.agents.isEmpty {
-                    Menu(store.selectedAgentName ?? l.text(.selectAgent)) {
-                        ForEach(selectedTask.agents, id: \.self) { agent in
-                            Button(agent) {
-                                Task { await store.selectAgent(agent) }
-                            }
-                        }
-                    }
-                    .menuStyle(.borderlessButton)
-                }
-            }
-
-            inspectorMetadata
-            if store.selectedIssue == nil {
-                inspectorTabBar
-                Group {
-                    switch store.inspectorTab {
-                    case .changes:
-                        ChangesView(language: l, patch: store.changes.patch, files: store.changes.changedFiles)
-                    case .files:
-                        FilesView(language: l, nodes: store.files)
-                    case .ports:
-                        PortsView(language: l, ports: store.ports) { port in
-                            store.openPort(port)
-                        }
-                    case .browser:
-                        BrowserView(language: l, url: store.browserURL)
-                    }
-                }
-                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .top)
-            }
-        }
-    }
-
-    private var inspectorTabBar: some View {
-        HStack(spacing: 8) {
-            ForEach(InspectorTab.allCases) { tab in
-                Button {
-                    store.inspectorTab = tab
-                } label: {
-                    HStack(spacing: 7) {
-                        Image(systemName: inspectorIcon(for: tab))
-                        Text(l.inspectorTab(tab))
-                    }
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundStyle(ShellPalette.text)
-                    .padding(.horizontal, 10)
-                    .padding(.vertical, 9)
-                    .frame(maxWidth: .infinity)
-                    .background(store.inspectorTab == tab ? ShellPalette.panelRaised : ShellPalette.panelAlt)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 10, style: .continuous)
-                            .stroke(store.inspectorTab == tab ? ShellPalette.lineStrong : ShellPalette.line, lineWidth: 1)
-                    )
-                    .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                }
-                .buttonStyle(.plain)
-            }
-        }
-    }
-
-    private var inspectorMetadata: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            if let issue = store.selectedIssue {
-                metadataRow(label: l("Issue Status", "이슈 상태"), value: l.status(issue.status))
-                metadataRow(label: l("Issue Kind", "이슈 종류"), value: userFacingIssueKind(issue.kind, language: l))
-                metadataRow(label: l("Risk", "리스크"), value: issue.riskLevel)
-                if let assignee = store.selectedIssueAssignee {
-                    metadataRow(label: l("Assignee", "담당"), value: "\(assignee.roleName) · \(assignee.executionAgentName)")
-                }
-                if !issue.acceptanceCriteria.isEmpty {
-                    metadataRow(label: l("Acceptance", "수용 기준"), value: issue.acceptanceCriteria.joined(separator: " · "))
-                }
-            }
-
-            if let reviewItem = store.selectedReviewQueueItem {
-                metadataRow(label: l("Review Queue", "리뷰 큐"), value: l.status(reviewItem.status))
-                if let branchName = reviewItem.branchName, !branchName.isEmpty {
-                    metadataRow(label: l.text(.branch), value: branchName)
-                }
-                if let worktreePath = reviewItem.worktreePath, !worktreePath.isEmpty {
-                    metadataRow(label: l.text(.worktree), value: worktreePath)
-                }
-                if let pullRequestUrl = reviewItem.pullRequestUrl, !pullRequestUrl.isEmpty {
-                    metadataRow(label: l.text(.pullRequest), value: pullRequestUrl)
-                }
-                if let pullRequestState = reviewItem.pullRequestState, !pullRequestState.isEmpty {
-                    metadataRow(label: l("PR State", "PR 상태"), value: l.status(pullRequestState))
-                }
-                if let mergeability = reviewItem.mergeability, !mergeability.isEmpty {
-                    metadataRow(label: l("Mergeability", "머지 가능성"), value: mergeability)
-                }
-                if let checksSummary = reviewItem.checksSummary, !checksSummary.isEmpty {
-                    metadataRow(label: l("Checks", "체크"), value: checksSummary)
-                }
-                if !reviewItem.requestedReviewers.isEmpty {
-                    metadataRow(label: l("Reviewers", "리뷰어"), value: reviewItem.requestedReviewers.joined(separator: ", "))
-                }
-                if let qaVerdict = reviewItem.qaVerdict, !qaVerdict.isEmpty {
-                    metadataRow(label: l("QA Verdict", "QA 판정"), value: l.status(qaVerdict))
-                }
-                if let qaFeedback = reviewItem.qaFeedback, !qaFeedback.isEmpty {
-                    metadataRow(label: l("QA Feedback", "QA 피드백"), value: qaFeedback)
-                }
-                if let ceoVerdict = reviewItem.ceoVerdict, !ceoVerdict.isEmpty {
-                    metadataRow(label: l("CEO Verdict", "CEO 판정"), value: l.status(ceoVerdict))
-                }
-                if let ceoFeedback = reviewItem.ceoFeedback, !ceoFeedback.isEmpty {
-                    metadataRow(label: l("CEO Feedback", "CEO 피드백"), value: ceoFeedback)
-                }
-                if let mergeCommitSha = reviewItem.mergeCommitSha, !mergeCommitSha.isEmpty {
-                    metadataRow(label: l("Merge Commit", "머지 커밋"), value: mergeCommitSha)
-                }
-            }
-
-            if let run = store.selectedRun {
-                metadataRow(label: l.text(.branch), value: run.branchName)
-                metadataRow(label: l.text(.base), value: run.baseBranch)
-                metadataRow(label: l.text(.worktree), value: run.worktreePath)
-                if let publish = run.publish {
-                    if let commitSha = publish.commitSha, !commitSha.isEmpty {
-                        metadataRow(label: l.text(.commit), value: commitSha)
-                    }
-                    if let pushedBranch = publish.pushedBranch, !pushedBranch.isEmpty {
-                        metadataRow(label: l.text(.pushedBranch), value: pushedBranch)
-                    }
-                    if let pullRequestUrl = publish.pullRequestUrl, !pullRequestUrl.isEmpty {
-                        metadataRow(label: l.text(.pullRequest), value: pullRequestUrl)
-                    }
-                    if let pullRequestState = publish.pullRequestState, !pullRequestState.isEmpty {
-                        metadataRow(label: l("PR State", "PR 상태"), value: l.status(pullRequestState))
-                    }
-                    if let reviewState = publish.reviewState, !reviewState.isEmpty {
-                        metadataRow(label: l("Publish Review", "퍼블리시 리뷰"), value: reviewState)
-                    }
-                    if let checksSummary = publish.checksSummary, !checksSummary.isEmpty {
-                        metadataRow(label: l("Publish Checks", "퍼블리시 체크"), value: checksSummary)
-                    }
-                    if let mergeability = publish.mergeability, !mergeability.isEmpty {
-                        metadataRow(label: l("Publish Mergeability", "퍼블리시 머지 가능성"), value: mergeability)
-                    }
-                    if let publishError = publish.error, !publishError.isEmpty {
-                        metadataRow(label: l.text(.publishError), value: publishError)
-                    }
-                }
-            } else if store.selectedIssue == nil {
-                Text(l.text(.selectTaskAndAgent))
-                    .font(.system(size: 12, weight: .medium))
-                    .foregroundStyle(ShellPalette.muted)
-            }
-        }
-        .padding(14)
-        .shellInset()
-    }
-
-    private func metadataRow(label: String, value: String) -> some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text(label.uppercased())
-                .font(.system(size: 9, weight: .bold))
-                .tracking(0.7)
-                .foregroundStyle(ShellPalette.faint)
-            Text(value)
-                .font(.system(size: 11, weight: .medium, design: .monospaced))
-                .foregroundStyle(ShellPalette.text)
-                .lineLimit(2)
-                .textSelection(.enabled)
-        }
-    }
-
-    private func inspectorIcon(for tab: InspectorTab) -> String {
-        switch tab {
-        case .changes:
-            return "arrow.triangle.branch"
-        case .files:
-            return "doc.text"
-        case .ports:
-            return "plugs.connected"
-        case .browser:
-            return "globe"
-        }
-    }
-}
-
-private struct InspectorContainerModifier: ViewModifier {
-    let embedded: Bool
-
-    func body(content: Content) -> some View {
-        if embedded {
-            content
-                .padding(14)
-                .background(
-                    RoundedRectangle(cornerRadius: ShellMetrics.radiusMedium, style: .continuous)
-                        .fill(ShellPalette.panel)
-                )
-                .overlay(
-                    RoundedRectangle(cornerRadius: ShellMetrics.radiusMedium, style: .continuous)
-                        .stroke(ShellPalette.line, lineWidth: 1)
-                )
-        } else {
-            content.shellCard()
-        }
-    }
-}
-
-private struct ChangesView: View {
-    let language: AppLanguage
-    let patch: String
-    let files: [String]
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            if files.isEmpty, patch.isEmpty {
-                EmptyStateView(
-                    image: "arrow.triangle.branch",
-                    title: language.text(.noChanges),
-                    subtitle: language.text(.noChangesSubtitle)
-                )
-            } else {
-                if !files.isEmpty {
-                    ScrollView(.horizontal, showsIndicators: false) {
-                        HStack(spacing: 8) {
-                            ForEach(files, id: \.self) { file in
-                                ShellTag(text: file, tint: ShellPalette.accent)
-                            }
-                        }
-                    }
-                }
-
-                ScrollView {
-                    Text(patch)
-                        .font(.system(size: 11, weight: .medium, design: .monospaced))
-                        .foregroundStyle(ShellPalette.text)
-                        .textSelection(.enabled)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(14)
-                }
-                .shellInset()
-            }
-        }
-    }
-}
-
-private struct FilesView: View {
-    let language: AppLanguage
-    let nodes: [FileTreeNodePayload]
-
-    var body: some View {
-        if nodes.isEmpty {
-            EmptyStateView(
-                image: "doc.text.magnifyingglass",
-                title: language.text(.noFileTree),
-                subtitle: language.text(.noFileTreeSubtitle)
-            )
-        } else {
-            ScrollView {
-                VStack(alignment: .leading, spacing: 0) {
-                    OutlineGroup(nodes, children: \.optionalChildren) { node in
-                        HStack(spacing: 10) {
-                            Image(systemName: node.isDirectory ? "folder.fill" : "doc.text")
-                                .foregroundStyle(node.isDirectory ? ShellPalette.accentWarm : ShellPalette.accent)
-                            Text(node.name)
-                                .font(.system(size: 12, weight: .medium))
-                                .foregroundStyle(ShellPalette.text)
-                            Spacer()
-                            if let size = node.sizeBytes, !node.isDirectory {
-                                Text(ByteCountFormatter.string(fromByteCount: size, countStyle: .file))
-                                    .font(.system(size: 10, weight: .medium))
-                                    .foregroundStyle(ShellPalette.muted)
-                            }
-                        }
-                        .padding(.vertical, 6)
-                    }
-                }
-                .padding(14)
-            }
-            .shellInset()
-        }
-    }
-}
-
-private struct PortsView: View {
-    let language: AppLanguage
-    let ports: [PortEntryPayload]
-    let onOpen: (PortEntryPayload) -> Void
-
-    var body: some View {
-        if ports.isEmpty {
-            EmptyStateView(
-                image: "plugs.connected",
-                title: language.text(.noPorts),
-                subtitle: language.text(.noPortsSubtitle)
-            )
-        } else {
-            ScrollView {
-                VStack(spacing: 10) {
-                    ForEach(ports) { port in
-                        HStack(alignment: .top, spacing: 12) {
-                            VStack(alignment: .leading, spacing: 5) {
-                                Text(port.label)
-                                    .font(.system(size: 13, weight: .semibold))
-                                    .foregroundStyle(ShellPalette.text)
-                                Text(port.url)
-                                    .font(.system(size: 11, weight: .medium, design: .monospaced))
-                                    .foregroundStyle(ShellPalette.muted)
-                            }
-
-                            Spacer()
-
-                            Button(language.text(.open)) {
-                                onOpen(port)
-                            }
-                            .buttonStyle(ShellTopBarButtonStyle(prominent: true))
-                        }
-                        .padding(14)
-                        .shellInset()
-                    }
-                }
-            }
-        }
-    }
-}
-
-private struct BrowserView: View {
-    let language: AppLanguage
-    let url: URL?
-
-    var body: some View {
-        if let url {
-            WebView(url: url)
-                .shellInset()
-        } else {
-            EmptyStateView(
-                image: "globe",
-                title: language.text(.browserIdle),
-                subtitle: language.text(.browserIdleSubtitle)
-            )
-        }
     }
 }
 
@@ -9072,7 +8414,7 @@ private struct OrgProfileBatchEditSheet: View {
     }
 }
 
-private struct EmptyStateView: View {
+struct EmptyStateView: View {
     let image: String
     let title: String
     let subtitle: String
@@ -9218,17 +8560,6 @@ private struct DesktopCommandMenu: Commands {
             }
             .keyboardShortcut("b", modifiers: [.command])
         }
-    }
-}
-
-private func localizedSourceKind(_ sourceKind: String, language: AppLanguage) -> String {
-    switch sourceKind.uppercased() {
-    case "LOCAL":
-        return language("Local", "로컬")
-    case "CLONED":
-        return language("Cloned", "복제됨")
-    default:
-        return sourceKind
     }
 }
