+++ src/main/kotlin/com/cotor/app/AppServer.kt
@@ -563,6 +563,13 @@ internal fun Application.cotorAppModule(
                 }
             }
 
+            route("/direct-chat") {
+                get("/providers") {
+                    if (!requireToken(token)) return@get
+                    call.respond(desktopService.directChatProviderCatalog())
+                }
+            }
+
             route("/skills") {
                 get {
                     if (!requireToken(token)) return@get
@@ -1945,11 +1952,17 @@ internal fun Application.cotorAppModule(
                         if (!requireToken(token)) return@get
                         val companyId = call.parameters["companyId"]
                             ?: return@get call.respond(HttpStatusCode.BadRequest, mapOf("error" to "companyId is required"))
+                        val cursor = call.request.queryParameters["cursor"]
+                            ?: call.request.header("Last-Event-ID")
                         call.respondTextWriter(ContentType.parse("application/x-ndjson")) {
                             writeCompanyEventStream(
-                                events = desktopService.companyEvents(companyId),
+                                events = desktopService.companyEvents(companyId, cursor),
                                 companyId = companyId,
-                                streamJson = streamJson
+                                streamJson = streamJson,
+                                initialCursor = cursor,
+                                gapResolver = { afterSequence ->
+                                    desktopService.companyEventGapSnapshot(companyId, afterSequence)
+                                }
                             )
                         }
                     }
@@ -2845,9 +2858,12 @@ private suspend fun RoutingContext.respondDesktopRequest(block: suspend () -> An
 private suspend fun Writer.writeCompanyEventStream(
     events: Flow<CompanyEventEnvelope>,
     companyId: String,
-    streamJson: Json
+    streamJson: Json,
+    initialCursor: String? = null,
+    gapResolver: suspend (Long?) -> CompanyEventEnvelope? = { null }
 ) = coroutineScope {
     val channel = events.produceIn(this)
+    var lastSequence = initialCursor?.toLongOrNull()
     try {
         while (currentCoroutineContext().isActive) {
             val next = withTimeoutOrNull(COMPANY_EVENT_STREAM_HEARTBEAT_MS) {
@@ -2864,7 +2880,9 @@ private suspend fun Writer.writeCompanyEventStream(
                                 type = "stream.heartbeat",
                                 title = "Stream heartbeat",
                                 createdAt = System.currentTimeMillis()
-                            )
+                            ),
+                            sequence = lastSequence,
+                            cursor = lastSequence?.toString()
                         )
                     )
                 )
@@ -2873,9 +2891,25 @@ private suspend fun Writer.writeCompanyEventStream(
                 continue
             }
             val envelope = next.getOrNull() ?: break
+            val sequence = envelope.sequence
+            if (sequence != null && lastSequence != null && sequence > lastSequence + 1) {
+                val gapSnapshot = gapResolver(lastSequence)
+                if (gapSnapshot != null) {
+                    write(streamJson.encodeToString(CompanyEventEnvelope.serializer(), gapSnapshot))
+                    write("\n")
+                    flush()
+                    lastSequence = gapSnapshot.sequence ?: lastSequence
+                    if (sequence <= (lastSequence ?: Long.MIN_VALUE)) {
+                        continue
+                    }
+                }
+            }
             write(streamJson.encodeToString(CompanyEventEnvelope.serializer(), envelope))
             write("\n")
             flush()
+            if (sequence != null) {
+                lastSequence = sequence
+            }
         }
     } finally {
         channel.cancel()
