+++ src/main/kotlin/com/cotor/app/LocalPlaywrightMarketingBrowserRunner.kt
@@ -8,12 +8,13 @@ import java.nio.file.Files
 import java.nio.file.Path
 import kotlin.io.path.createDirectories
 import kotlin.io.path.deleteIfExists
-import kotlin.io.path.exists
 
 class LocalPlaywrightMarketingBrowserRunner internal constructor(
     private val appHomeProvider: () -> Path = { defaultDesktopAppHome() },
     private val commandAvailability: (String) -> Boolean = ::marketingCommandAvailable,
-    private val processRunner: LocalSkillProcessRunner = defaultLocalSkillProcessRunner
+    private val processRunner: LocalSkillProcessRunner = defaultLocalSkillProcessRunner,
+    private val playwrightDependencyResolver: LocalPlaywrightDependencyResolver =
+        LocalPlaywrightDependencyResolver(commandAvailability, processRunner)
 ) : MarketingBrowserRunner {
     private val json = Json {
         ignoreUnknownKeys = true
@@ -36,17 +37,17 @@ class LocalPlaywrightMarketingBrowserRunner internal constructor(
                 MARKETING_BROWSER_MAX_RUNTIME_SECONDS
             )
         )
-        require(commandAvailability("node") && commandAvailability("npm")) {
-            "Marketing browser execution requires node and npm so Playwright can run without storing browser credentials in Cotor."
-        }
         val runtimeDir = runtimeRoot
             .resolve("marketing-browser")
         val inputDir = runtimeDir.resolve("inputs")
         val scriptPath = runtimeDir.resolve("marketing-runner.js")
         runtimeDir.createDirectories()
         inputDir.createDirectories()
-        // Install phase runs with its own 120s minimum and must not be capped by the run-phase timeout.
-        ensurePlaywrightDependency(runtimeDir, normalizedCommand.maxRuntimeSeconds)
+        val nodePath = playwrightDependencyResolver.requireNodePath(
+            runtimeDir = runtimeDir,
+            timeoutSeconds = normalizedCommand.maxRuntimeSeconds,
+            label = "Marketing browser execution"
+        )
         writeTextAtomically(scriptPath, playwrightRunnerScript)
         val inputPath = Files.createTempFile(inputDir, "marketing-command-", ".json")
         try {
@@ -57,7 +58,7 @@ class LocalPlaywrightMarketingBrowserRunner internal constructor(
                 workingDirectory = runtimeDir,
                 timeoutSeconds = timeoutSeconds,
                 timeoutMessage = "Marketing browser execution timed out after ${timeoutSeconds}s.",
-                environment = mapOf("NODE_PATH" to runtimeDir.resolve("node_modules").toString())
+                environment = mapOf("NODE_PATH" to nodePath.toString())
             )
             val output = result.output.trim()
             if (result.exitCode != 0) {
@@ -72,36 +73,6 @@ class LocalPlaywrightMarketingBrowserRunner internal constructor(
             }
         } finally {
             inputPath.deleteIfExists()
-        }
-    }
-
-    private suspend fun ensurePlaywrightDependency(runtimeDir: Path, timeoutSeconds: Int) {
-        withLocalSkillRuntimeMutationLock(runtimeDir) {
-            val packageDir = runtimeDir.resolve("node_modules").resolve("playwright")
-            if (packageDir.exists()) {
-                return@withLocalSkillRuntimeMutationLock
-            }
-            val installTimeoutSeconds = timeoutSeconds.coerceAtLeast(120).toLong()
-            val result = processRunner.run(
-                command = listOf(
-                    "npm",
-                    "install",
-                    "--silent",
-                    "--no-audit",
-                    "--no-fund",
-                    "--prefix",
-                    runtimeDir.toString(),
-                    "playwright"
-                ),
-                workingDirectory = runtimeDir,
-                timeoutSeconds = installTimeoutSeconds,
-                timeoutMessage = "Playwright dependency install timed out after ${installTimeoutSeconds}s.",
-                environment = emptyMap()
-            )
-            val output = result.output.trim()
-            if (result.exitCode != 0) {
-                error(output.ifBlank { "Playwright dependency install failed with exit ${result.exitCode}." })
-            }
         }
     }
